Initial working character level transcription extraction and loading
Modify unknown char mapping behavior

Absorb length information into char_map::decoded

Wow, passes tests

Loader example now works (was missing add_providers after creation)

Add stub file for audio provider

Some additional notes

Flesh out audio provider a bit more

+++ loader/src/buffer_in.cpp
@@ -26,7 +26,7 @@
 
 using namespace std;
 
-buffer_in::buffer_in(int size) {
+buffer_in::buffer_in(size_t size) {
 }
 
 buffer_in::~buffer_in() {
