+++ loader/src/buffer_pool_in.cpp
@@ -30,17 +30,16 @@
 
 using namespace std;
 
-buffer_pool_in::buffer_pool_in(const std::vector<uint32_t>& initial_sizes)
-    : buffer_pool() {
+buffer_pool_in::buffer_pool_in(const std::vector<size_t>& initial_sizes)
+:buffer_pool()
+{
     for (int i = 0; i < _count; i++) {
         _bufs.push_back(make_shared<buffer_in_array>(initial_sizes));
     }
 
 }
 
-buffer_pool_in::~buffer_pool_in() {
-
-}
+buffer_pool_in::~buffer_pool_in() {}
 
 buffer_in_array& buffer_pool_in::getForWrite()
 {
@@ -49,63 +48,72 @@ buffer_in_array& buffer_pool_in::getForWrite()
         b->reset();
     }
     return buf_ary;
-    // _bufs[_writePos][0]->reset();
-    // _bufs[_writePos][1]->reset();
-    // return _bufs[_writePos];
 }
 
-buffer_in_array& buffer_pool_in::getForRead() {
+buffer_in_array& buffer_pool_in::getForRead()
+{
     reraiseException();
     return *_bufs[_readPos];
 }
 
-buffer_in_array& buffer_pool_in::getPair(int bufIdx) {
+buffer_in_array& buffer_pool_in::getPair(int bufIdx)
+{
     assert(bufIdx >= 0 && bufIdx < _count);
     return *_bufs[bufIdx];
 }
 
-void buffer_pool_in::advanceReadPos() {
+void buffer_pool_in::advanceReadPos()
+{
     _used--;
     advance(_readPos);
 }
 
-void buffer_pool_in::advanceWritePos() {
+void buffer_pool_in::advanceWritePos()
+{
     _used++;
     advance(_writePos);
     clearException();
 }
 
-bool buffer_pool_in::empty() {
+bool buffer_pool_in::empty()
+{
     assert(_used >= 0);
     return (_used == 0);
 }
 
-bool buffer_pool_in::full() {
+bool buffer_pool_in::full()
+{
     assert(_used <= _count);
     return (_used == _count);
 }
 
-std::mutex& buffer_pool_in::getMutex() {
+std::mutex& buffer_pool_in::getMutex()
+{
     return _mutex;
 }
 
-void buffer_pool_in::waitForNonEmpty(std::unique_lock<std::mutex>& lock) {
+void buffer_pool_in::waitForNonEmpty(std::unique_lock<std::mutex>& lock)
+{
     _nonEmpty.wait(lock);
 }
 
-void buffer_pool_in::waitForNonFull(std::unique_lock<std::mutex>& lock) {
+void buffer_pool_in::waitForNonFull(std::unique_lock<std::mutex>& lock)
+{
     _nonFull.wait(lock);
 }
 
-void buffer_pool_in::signalNonEmpty() {
+void buffer_pool_in::signalNonEmpty()
+{
     _nonEmpty.notify_all();
 }
 
-void buffer_pool_in::signalNonFull() {
+void buffer_pool_in::signalNonFull()
+{
     _nonFull.notify_all();
 }
 
-void buffer_pool_in::advance(int& index) {
+void buffer_pool_in::advance(int& index)
+{
     // increment index and reset to 0 when index hits `_count`
     if (++index == _count) {
         index = 0;
